+++ src/config/plugin.ts
@@ -1,8 +1,5 @@
-/* eslint-disable no-await-in-loop */
-import type {PackageInformation, PackageLocator, getPackageInformation} from 'pnpapi'
-
 import {sync} from 'globby'
-import {basename, dirname, join, parse, relative, sep} from 'node:path'
+import {join, parse, relative, sep} from 'node:path'
 import {inspect} from 'node:util'
 
 import {Command} from '../command'
@@ -14,10 +11,11 @@ import {Topic} from '../interfaces/topic'
 import {loadWithData, loadWithDataFromManifest} from '../module-loader'
 import {OCLIF_MARKER_OWNER, Performance} from '../performance'
 import {cacheCommand} from '../util/cache-command'
-import {readJson, requireJson, safeReadJson} from '../util/fs'
+import {findRoot} from '../util/find-root'
+import {readJson, requireJson} from '../util/fs'
 import {compact, isProd, mapValues} from '../util/util'
 import {tsPath} from './ts-node'
-import {Debug, getCommandIdPermutations, resolvePackage} from './util'
+import {Debug, getCommandIdPermutations} from './util'
 
 const _pjson = requireJson<PJSON>(__dirname, '..', '..', 'package.json')
 
@@ -32,157 +30,6 @@ function topicsToArray(input: any, base?: string): Topic[] {
     input[k].name = k
     return [{...input[k], name: `${base}${k}`}, ...topicsToArray(input[k].subtopics, `${base}${input[k].name}`)]
   })
-}
-
-// essentially just "cd .."
-function* up(from: string) {
-  while (dirname(from) !== from) {
-    yield from
-    from = dirname(from)
-  }
-
-  yield from
-}
-
-async function findPluginRoot(root: string, name?: string) {
-  // If we know the plugin name then we just need to traverse the file
-  // system until we find the directory that matches the plugin name.
-  if (name) {
-    for (const next of up(root)) {
-      if (next.endsWith(basename(name))) return next
-    }
-  }
-
-  // If there's no plugin name (typically just the root plugin), then we need
-  // to traverse the file system until we find a directory with a package.json
-  for (const next of up(root)) {
-    // Skip the bin directory
-    if (
-      basename(dirname(next)) === 'bin' &&
-      ['dev', 'dev.cmd', 'dev.js', 'run', 'run.cmd', 'run.js'].includes(basename(next))
-    ) {
-      continue
-    }
-
-    try {
-      const cur = join(next, 'package.json')
-      if (await safeReadJson<PJSON>(cur)) return dirname(cur)
-    } catch {}
-  }
-}
-
-/**
- * Find package root for packages installed into node_modules. This will go up directories
- * until it finds a node_modules directory with the plugin installed into it
- *
- * This is needed because some oclif plugins do not declare the `main` field in their package.json
- * https://github.com/oclif/config/pull/289#issuecomment-983904051
- *
- * @returns string
- * @param name string
- * @param root string
- */
-async function findRootLegacy(name: string | undefined, root: string): Promise<string | undefined> {
-  for (const next of up(root)) {
-    let cur
-    if (name) {
-      cur = join(next, 'node_modules', name, 'package.json')
-      if (await safeReadJson<PJSON>(cur)) return dirname(cur)
-
-      const pkg = await safeReadJson<PJSON>(join(next, 'package.json'))
-      if (pkg?.name === name) return next
-    } else {
-      cur = join(next, 'package.json')
-      if (await safeReadJson<PJSON>(cur)) return dirname(cur)
-    }
-  }
-}
-
-let pnp: {
-  getPackageInformation: typeof getPackageInformation
-  getLocator: (name: string, reference: string | [string, string]) => PackageLocator
-  getDependencyTreeRoots: () => PackageLocator[]
-}
-
-function maybeRequirePnpApi(root: string): unknown {
-  if (pnp) return pnp
-  // Require pnpapi directly from the plugin.
-  // The pnpapi module is only available if running in a pnp environment.
-  // Because of that we have to require it from the root.
-  // Solution taken from here: https://github.com/yarnpkg/berry/issues/1467#issuecomment-642869600
-  // eslint-disable-next-line node/no-missing-require
-  pnp = require(require.resolve('pnpapi', {paths: [root]}))
-
-  return pnp
-}
-
-const getKey = (locator: PackageLocator | string | [string, string] | undefined) => JSON.stringify(locator)
-const isPeerDependency = (
-  pkg: PackageInformation | undefined,
-  parentPkg: PackageInformation | undefined,
-  name: string,
-) => getKey(pkg?.packageDependencies.get(name)) === getKey(parentPkg?.packageDependencies.get(name))
-
-/**
- * Traverse PnP dependency tree to find package root
- *
- * Implementation taken from https://yarnpkg.com/advanced/pnpapi#traversing-the-dependency-tree
- */
-function findPnpRoot(name: string, root: string): string | undefined {
-  maybeRequirePnpApi(root)
-  const seen = new Set()
-
-  const traverseDependencyTree = (locator: PackageLocator, parentPkg?: PackageInformation): string | undefined => {
-    // Prevent infinite recursion when A depends on B which depends on A
-    const key = getKey(locator)
-    if (seen.has(key)) return
-
-    const pkg = pnp.getPackageInformation(locator)
-
-    if (locator.name === name) {
-      return pkg.packageLocation
-    }
-
-    seen.add(key)
-
-    for (const [name, referencish] of pkg.packageDependencies) {
-      // Unmet peer dependencies
-      if (referencish === null) continue
-
-      // Avoid iterating on peer dependencies - very expensive
-      if (parentPkg !== null && isPeerDependency(pkg, parentPkg, name)) continue
-
-      const childLocator = pnp.getLocator(name, referencish)
-      const foundSomething = traverseDependencyTree(childLocator, pkg)
-      if (foundSomething) return foundSomething
-    }
-
-    // Important: This `delete` here causes the traversal to go over nodes even
-    // if they have already been traversed in another branch. If you don't need
-    // that, remove this line for a hefty speed increase.
-    seen.delete(key)
-  }
-
-  // Iterate on each workspace
-  for (const locator of pnp.getDependencyTreeRoots()) {
-    const foundSomething = traverseDependencyTree(locator)
-    if (foundSomething) return foundSomething
-  }
-}
-
-async function findRoot(name: string | undefined, root: string) {
-  if (name) {
-    let pkgPath
-    try {
-      pkgPath = resolvePackage(name, {paths: [root]})
-    } catch {}
-
-    if (pkgPath) return findPluginRoot(dirname(pkgPath), name)
-
-    return process.versions.pnp ? findPnpRoot(name, root) : findRootLegacy(name, root)
-  }
-
-  return findPluginRoot(root)
 }
 
 const cachedCommandCanBeUsed = (manifest: Manifest | undefined, id: string): boolean =>
